First commit. Added README.md and prog_struct.txt.
Defined the limits of this program and what it can and cannot be used for
in the readme. I tried to give the reader a clear understanding of the
capabilities.

I started to build out the structure of the program. Where certain
functions will go and I have started thinking about how the different
modules will interact with each other.

I have also considered that some of the basic functionality, calculating
the centroid, Ixx, Iyy, J, etc. could be used for a pile group analysis or
beam group analysis program. I am considering breaking this out into a
more generic python package. I don't want to get to into the weeds and
stay focused on what needs to get done and the code can be made more
generic, and refactored later.