Wrote test functions for helper functions. Tested helper functions.
Tested all helper functions written to this point. Had to add to the
existing bolt and force data structures to get some of the testing to work
out. Needed a place to store the local coordinate points because the user
coordinate points are by design an immutable tuple.

The polar moment of inertia, calc_j, equation was incorrect. Just need to
add ixx and iyy instead of squaring them and adding them. Corrected now.