Documented data structure parameters and location.
I noticed this was going to become a source of confusion for myself and
possibly others in the future. So I decided to document it somewhere. I
wasn't really sure where to document it and it may move in the future. It
is also subject to change but I think I have all the parameters I need for
this module.