Added eccentric shear in the plane elastic function.
Added function and documentation. Working on testing. It is complicated
testing. I feel the need to check that the magnitude and direction of the
shear on a bolt in each quadrant around a load application point is
correct. Also that a load application point is checked in each quandrant.
This requires a fair bit of manual testing.