Updated bolt reaction coordinate system.
Instead of outputing the forces acting on the bolts I have switched to
outputing the reactions. I.e. the force and direction required at the
bolt locations to resist the applied force.

This matches the coordinate system used in the paper that will be followed
implement the eccentric shear in plane plastic analysis. Instead of trying
to convert between the two systems, just using the one that has reactions,
which is really what I should be doing anyways.