Refactored some functions. Started on eccentric shear in plane by plastic.
I decided that the functions that add to the existing bolt and force data
structures should only be called once by an outside program. I had to edit
other functions that make calls to make sure the data they need to
populated. Eventually I will add exceptions to make sure the data is there
before subsequent functions are ran.

Started working on eccentric shear in the plane by the elastic method. It
is involved and I am having to add a number of helper function that were
not in the original program structure. I am following a paper by Brandt to
implement his solution for iterating to the correct plastic instanteous
center.