Updated unit testing. The in plane test was broken into two files.
The elastic in plane shear testing was becoming unwieldy. I think this is
just the nature of testing a program like this and, well, using a large
problem to test it.

Also, refactored the test_local_bolt_coords so that the correct answer is
just included in the code instead of opening an outside file. It is just
cleaner, more compact, and less prone to problems this way.

Started on plastic in plane shear testing.