Add new data members to the data structure, Rsx and Rsy.
I was having trouble coming up with a good way to combine the results of
different effects. For example direct shear reactions and elastic
eccentric force in the plane. I decided it would be easiest to add data
members for the results of each analysis type. This way the results stay
coupled with each bolt, all the results can be printed for the user, and
the results can be easily and logically combined.