Refactoring code and data structures.
The alogrithm presented by Brandt probably works great for determining
the bolt group coefficients in an efficient manner. However, I want to be
able to look at the force on each bolt based on the actual force input and
the actual capacity. Brandt uses ratios for all of these things. He also
never checks that the bolt group is rotationally static that I can tell. I
think there should be an explicit check for this.

Writing something that I think will use his alogrithm for the next guess
at the IC but calculates the actual forces.