Figured out how to use Brandt's algorithm and calc reactions of bolts.
Brandt's algorithm works really well as to be expected. I was really
confused about how to arrive at bolt forces. It is as simple as
calculating everything and using the final run of the interation as the
results. The results have the most accurate approximation of the
deflection of each bolt, use that in the constitutive equation to back out
the force in each bolt. Another check that the applied load from summing
forces in the x, y, and moments can be used here.

Most of the code to this point is correct.

Updated testing code to reflect some changes in the main code body. Have
not run tests.

Need to figure out how to calculate the perpendicular distance from the
line of action of the force to the IC.