Changed variable names. Corrected calculation of force eccentricity.
Changed local variable names so the entire program read more clearly and
hopefully stave off confusion in the future. I tried to follow a
reasonable naming convention and may need to put together a FBD and a few
pictorials to include in the documentation.

Added the variable 'ec' to represent the distance from the line of action of
the force to the centroid. I was mistakenly calculating this as the
distance from the centroid to the point of application of the force and
these two distances, in general, are not equal.

Updated the calculation of 'r', the distance from the line of action of
the force to the IC. I was mistakenly using calculating this as the
distance from the IC to the point of application of the force. These two
distances, in general, are not equal.