Added documentation. Updated calc_bolt_reactions. Split calc_r.
Continued adding documentation to the plastic functions. There is so much
good work done, I don't want to get distracted and come back to this and
have to figure it all out again.

Updated teh calc_bolt_reations functions. Noticed that it wasn't using the
correct variable name. This function is untested.

Split calc_r into calc_force_location_wrt_centroid and calc_r. When
writing the documentation for the original calc_r noticed that there was
no good way to describe the function because it was really doing two
unrelated operations. Broke into two functions. Need to update the rest of
the code to reflect this change.