All tests passed. Need test for calc_bolt_reactions.
When calc_r was split the function name for the new function was incorrect
it should be calc_force_location_wrt_ic, which is what it actually does.
Also, there is already a function with the
calc_force_location_wrt_centroid that does exactly that. Fixed this and
associated problems in program and testing code.

Updated test code to incorporate the split in the calc_r function.

Determined that the calc_r function isn't really needed but leaving it
because it may be needed in the future.

Refactored iterate_to_ic. The way the initialization was written and the
the code in the while loop made it difficult to break at a logical point
without some work around code; mainly assigning the location of the IC to
a temp variable at the beginning of each loop...it just didn't make any
sense. Refactored the whole function and now I think it is a much more
logical flow.

Added functionality to calculate and return Ce and Cu from the
iterate_to_ic function.

Updated test code to incorporate the name change to iterate_to_ic.

Changed calc_d to calc_bolt_location_wrt_ic.

Most of the documentation is complete.